Remove C-h binding, comment out magit-fullscreen
Removing the C-h -> backspace binding, not used much with the new keyboard
layout.

Commenting out the magit-fullscreen stuff to try magit without it again.  Main
thing was annoyance with magit buffers hanging around, but with midnight-mode
this shouldn't be as much of an issue...Leaving in place just in case I change
my mind again.